+++ c/containers/lhmsi.h
@@ -20,7 +20,7 @@
 typedef struct _lhmsie_t {
 	int   ideal_index;
 	char* key;
-	long long value;
+	int value;
 	char  free_flags;
 	struct _lhmsie_t *pprev;
 	struct _lhmsie_t *pnext;
@@ -43,8 +43,8 @@ lhmsi_t* lhmsi_alloc();
 lhmsi_t* lhmsi_copy(lhmsi_t* pmap);
 void  lhmsi_free(lhmsi_t* pmap);
 void  lhmsi_put(lhmsi_t* pmap, char* key, int value, char free_flags);
-long long lhmsi_get(lhmsi_t* pmap, char* key); // caller must do lhmsi_has_key to check validity
-int lhmsi_test_and_get(lhmsi_t* pmap, char* key, long long* pval); // *pval undefined if return is FALSE
+int lhmsi_get(lhmsi_t* pmap, char* key); // caller must do lhmsi_has_key to check validity
+int lhmsi_test_and_get(lhmsi_t* pmap, char* key, int* pval); // *pval undefined if return is FALSE
 lhmsie_t* lhmsi_get_entry(lhmsi_t* pmap, char* key);
 int   lhmsi_has_key(lhmsi_t* pmap, char* key);
